Fixed a casting error.

+++ src/sql.h
@@ -86,7 +86,7 @@ void zone_trophy_update();
 #define EXPLOG "exp"
 #define CONNECTLOG "connect"
 
-#define FRAGS_TO_LEVEL( frags ) ( (frags / .6) + 25 )
+#define FRAGS_TO_LEVEL( frags ) ( (int)(frags / .6) + 25 )
 #define LEVEL_TO_FRAGS( level ) ( (level <= 25) ? 0 : (( float )( level - 25 ) * .6) )
 
 void sql_log(P_char ch, char * kind, char * format, ...);
